new release 247.3-1
Closely tracks the latest Fedora changes up to 247.3. The primary
difference is that networkd and the standalone binaries are not split
out into their own subpackages. Didn't seem worthwhile to split out
networkd for a couple of MB and deal with the configuration changes
related to missing the service files installed. I excluded the
standalone binary subpackage since it's intended for non-systemd
systems.