249.4-2.13: Move to dist-git layout and add back removed files
To enable a more streamlined way of pulling in changes from rawhide,
let's work to reduce the diff between the c8s-hyperscale-sig branch
and the rawhide branch as much as possible. The first step is to stop
using the SOURCES/ and SPECS/ directories as those aren't used by the
rawhide branch. Instead, everything exists in the root repository
directory.

Additionally, this diff pulls in the .gitignore file from the rawhide
branch as our own .gitignore/ relied on the SOURCES/ directory which
means it didn't work anymore after moving the files out of that
directory.

To further reduce the diff with the main branch, let's also add back
the deleted files from the main branch . Note that while we add back
the files, we don't add them back to the spec itself, so they'll
exist but won't be used.

Adding the files back allows us to merge changes from the main branch
into the hyperscale branch without running into merge conflicts because
the file was modified in the main branch but we deleted it. It also
makes it easier at a glance to compare the diff from the hyperscale
branch with the main branch since the diff will be smaller.

The files were taken from commit 46a40810, which is the commit that
c8s-hyperscale-sig was based on for the 249-4 release.

We don't add back the sources file because then we'd need to modify it
to specify the current release which would result in a merge conflict
when merging changes from Fedora so we keep it deleted.