250.3-6.7: Backport some upstream fixes + move files to networkd
We backport two fixes that ensure systemd-tmpfiles doesn't fail
due to a missing systemd-network user. The upstream fixes move the
networkd related stuff in tmpfiles.d/systemd.conf to
tmpfiles.d/systemd-network.conf. We then move this new file to the
networkd package so that it is only applied when the networkd
package is installed which also ensures the systemd-network user
will exist.