Run commands in %posttrans script unconditionally
Because https://github.com/rpm-software-management/rpm/commit/3848c97cb227e7c018781aa7d5e1e46990ce1ffb
is not in c9s, we can't detect upgrades in %posttrans. Luckily, the
stuff we run in %posttrans is a noop in chroots and we never expect
to install systemd on a live system in the first place so this shouldn't
be a problem.

Even if systemd was installed on a live system, all errors from commands
in scriptlets are ignored so at worst there will be some errors but nothing
that will actually prevent installation of the rpm.