Add back %postun scriptlets to support downgrades
On downgrades, only %posttrans of the package we're downgrading to
will run. Because the previous version we're upgrading from doesn't
have a %posttrans scriptlet, the end result is that we don't restart
any daemons. Let's make sure we restart daemons on downgrades by adding
%postun scriptlets that are identical to the %posttrans scriptlets. These
will run when we're downgrading to an older version without %posttrans.

This means we'll do double restarts when upgrading from 256 => 256 but
we're already doing double restarts when upgrading from 255 => 256 so a
few more won't hurt.