systemd.255-1.5: Only run daemon-reexec once during posttrans
During systemd 255 -> 256 upgrade, we noticed systemd daemon-reexec
being run 5 times in a row during upgrade:
1. systemd 255 postun
2. systemd udev 255 postun
3. systemd container 255 postun
4. systemd 255 triggerun
5. systemd 256 posttrans

All but 5 is unnecessary. To do this, we create a maintenance release
for 255 with systemctl daemon-reexec only once during post transaction.
This is always run at the end of the transaction so all uninstall/install
scripts have already been run.