Remove restart-required and %postun logic
The current effect of the restart-required logic is that we only
execute %postun of the old package on upgrade and not %posttrans of
the new package. This is exactly what we want to avoid as deploying
fixes to %postun requires two full rollouts before they take effect
compared to %posttrans which takes effect immediately.

Because we've already done a release with the restart-required logic,
the net effect on upgrade/downgrade now will be the following:

- On upgrade, the restart-required file will not be created anymore,
which means that %postun of the old package will be skipped. %posttrans
of the new package will run.
- On downgrade, the restart-required file will be created, which means
the %posttrans of the old package will run. As a result, we don't need
the %postun scriptlets anymore in the new package and can remove the logic
and solely rely on %posttrans in the future.