restart networkd on update
Facebook-only change, restart networkd after upgrade.
Historically networkd was never restarted because it caused link flap
and IPv6 route deletion. With this issues fixed, and the new
configuration serialization/deserialization, we can now restart
networkd.
This solves a whole range of client-daemon communication issues when
the client binary is updated but the daemon isn't restarted.