releng: rework autorelease logic to be less strict
The old logic requires all scanned tags to be at the same version.
This is now too strict since we have stopped regularly publish releases
for --release=9/10 (we do --release=9_z/10_z). At the same time, it's
possible that we will come back to old setup at some point.

To take this into account, the new logic scans all available tags for
latest maxima version. If it matches one writted in systemd.spec - it's
good. We bump the version_override.

If found max version is lower - there is nothing to do. The version in
systemd.spec should just build.

If found max verion is higher - we abort.